backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
google-generativeai==0.3.2
atproto==0.0.45
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
psycopg2-binary==2.9.9

